requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-wtf==1.2.1
email-validator==2.1.0.post1
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.1
numpy>=1.24.4
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
geocoder==1.38.1
geoip2>=4.7.0
python-ipware==3.0.0
geopy>=2.4.0
user-agents>=2.2.0
